Moved the Serialization and Deserialization of Sharing Model from InputFileExtensions to SharingModel to as it is no longer dependent on the methods in InputFileExtensions.
