+++ Application/EdFi.Ods.AdminApp.Management/ClaimSetEditor/SharingModel.cs
@@ -4,7 +4,10 @@
 // See the LICENSE and NOTICES files in the project root for more information.
 
 using System.Collections.Generic;
+using System.IO;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using Newtonsoft.Json.Serialization;
 
 namespace EdFi.Ods.AdminApp.Management.ClaimSetEditor
 {
@@ -12,6 +15,31 @@ namespace EdFi.Ods.AdminApp.Management.ClaimSetEditor
     {
         public string Title { get; set; }
         public SharingTemplate Template { get; set; }
+
+        public static string SerializeFromSharingModel(SharingModel model)
+        {
+            return JsonConvert.SerializeObject(model, UsingIndentedCamelCase());
+        }
+
+        public static SharingModel DeserializeToSharingModel(Stream jsonStream)
+        {
+            using (var streamReader = new StreamReader(jsonStream))
+                using (JsonReader jsonReader = new JsonTextReader(streamReader))
+                {
+                    return JsonSerializer
+                        .Create(UsingIndentedCamelCase())
+                        .Deserialize<SharingModel>(jsonReader);
+                }
+        }
+
+        private static JsonSerializerSettings UsingIndentedCamelCase()
+        {
+            return new JsonSerializerSettings
+            {
+                ContractResolver = new CamelCasePropertyNamesContractResolver(),
+                Formatting = Formatting.Indented
+            };
+        }
     }
 
     public class SharingTemplate
